pyproject.toml
[tool.poetry]
name = "django-layout"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
loguru = "^0.7.2"
mysqlclient = "^2.2.0"
djangorestframework = "^3.14.0"
django-redis = "^5.4.0"
channels = {extras = ["daphne"], version = "^4.0.0"}
djangorestframework-simplejwt = "^5.3.0"
channels-redis = "^4.1.0"
drf-yasg2 = "^1.19.4"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
